Refactor now-playing code and add format flags.
